Move tab area to move suitable place
