+++ src/components/sidebar/linkView/multiTabTextarea.scss
